Move top-level conftest into api level
