+++ tests/functional/services/api/accounts/users/test_get.py
@@ -1,8 +1,8 @@
 import pytest
 
+from tests.functional.services.api.conftest import FT_ACCOUNT, get_ft_user
 from tests.functional.services.api.conftest import USER_API_CONFS
-from tests.functional.utils.http_utils import http_get, APIResponse
-from tests.functional.conftest import FT_ACCOUNT, get_ft_user
+from tests.functional.services.utils.http_utils import http_get, APIResponse
 
 
 @pytest.mark.parametrize('api_conf', USER_API_CONFS)
