+++ tests/functional/services/api/accounts/users/test_post.py
@@ -1,9 +1,9 @@
 import pytest
 
 from tests.functional.services.api.accounts.users import delete_ft_account_user
+from tests.functional.services.api.conftest import FT_ACCOUNT
 from tests.functional.services.api.conftest import USER_API_CONFS
-from tests.functional.utils.http_utils import http_post, APIResponse
-from tests.functional.conftest import FT_ACCOUNT
+from tests.functional.services.utils.http_utils import http_post, APIResponse
 
 
 @pytest.mark.parametrize('api_conf', USER_API_CONFS)
