+++ tests/functional/services/api/archives/conftest.py
@@ -1,10 +1,10 @@
 import pytest
 
 from tests.functional.services.api.conftest import USER_API_CONFS
+from tests.functional import get_logger
 from tests.functional.services.api.images import wait_for_image_to_analyze, get_image_id, \
     get_image_digest
-from tests.functional.utils.http_utils import http_post, RequestFailedError, http_del
-from tests.functional.conftest import get_logger
+from tests.functional.services.utils.http_utils import http_post, RequestFailedError, http_del
 
 _logger = get_logger(__name__)
