+++ tests/functional/services/api/images/by_id/test_get.py
@@ -3,7 +3,7 @@ import pytest
 from tests.functional.services.api.images import get_image_id, get_image_tag, \
     wait_for_image_to_analyze
 from tests.functional.services.api.policies import get_first_policy_id
-from tests.functional.utils.http_utils import http_get, APIResponse
+from tests.functional.services.utils.http_utils import http_get, APIResponse
 
 
 class TestImagesByIDAPIGetReturns200:
