+++ tests/functional/services/api/repositories/test_post.py
@@ -1,7 +1,7 @@
 import pytest
 
 from tests.functional.services.api.conftest import USER_API_CONFS
-from tests.functional.utils.http_utils import http_post, APIResponse
+from tests.functional.services.utils.http_utils import http_post, APIResponse
 
 
 @pytest.mark.parametrize('api_conf', USER_API_CONFS)
