+++ tests/functional/services/api/summaries/test_get.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.functional.utils.http_utils import http_get, APIResponse
+from tests.functional.services.utils.http_utils import http_get, APIResponse
 
 
 class TestSummariesAPIGetReturns200:
