+++ tests/functional/services/api/system/test_delete.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.functional.utils.http_utils import http_get, APIResponse, http_put, http_del, \
+from tests.functional.services.utils.http_utils import http_get, APIResponse, http_put, http_del, \
     RequestFailedError
