+++ tests/functional/services/api/system/test_get.py
@@ -1,7 +1,7 @@
 import pytest
 
 from tests.functional.services.api.conftest import USER_API_CONFS
-from tests.functional.utils.http_utils import http_get, APIResponse
+from tests.functional.services.utils.http_utils import http_get, APIResponse
 
 
 @pytest.mark.parametrize('api_conf', USER_API_CONFS)
