+++ tests/functional/services/api/system/test_post.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.functional.utils.http_utils import http_post, APIResponse
+from tests.functional.services.utils.http_utils import http_post, APIResponse
 
 
 class TestSystemAPIPostReturns200:
