+++ tests/functional/services/api/test_default_get.py
@@ -1,4 +1,4 @@
-from tests.functional.utils.http_utils import http_get, APIResponse
+from tests.functional.services.utils.http_utils import http_get, APIResponse
 
 
 class TestDefaultAPIGetReturns200:
@@ -6,6 +6,7 @@ class TestDefaultAPIGetReturns200:
     This is for GET methods on endpoints under the "Default" section of the Swagger Spec
     Note: only run as admin right now
     """
+
     def test_root(self):
         resp = http_get([])
         assert resp == APIResponse(200)
